+++ tests/testthat/test-fit_model.R
@@ -26,40 +26,12 @@ test_that("Bad params w/ failing fit issues warning and returns NA", {
   # the Stan sampler but it terminates unexpectedly with an error, which is the
   # desired testing condition.
 
+  # Data loaded in from setup.R
   # Parameters
   parameters <- list(
     generation_interval = sir_gt_pmf,
     delay_interval = NA,
     right_truncation = NA
-  )
-  # Data -- 5 points only
-  data_path <- test_path("data", "test_data.parquet")
-  con <- DBI::dbConnect(duckdb::duckdb())
-  data <- DBI::dbGetQuery(
-    con,
-    "
-                         SELECT
-                           report_date,
-                           reference_date,
-                           disease,
-                           geo_value AS state_abb,
-                           value AS confirm
-                         FROM read_parquet(?)
-                         ORDER BY reference_date
-                         LIMIT 5
-                          ",
-    params = list(data_path)
-  )
-  DBI::dbDisconnect(con)
-  # Priors
-  priors <- list(
-    rt = list(
-      mean = 1,
-      sd = 0.2
-    ),
-    gp = list(
-      alpha_sd = 0.05
-    )
   )
   # Sampler
   sampler_opts <- list(